Add placeholder URLs and split up views
The urlconf now has sections for the different views we want. Test
pages are gone, and the non-text-generating views are now an index
and an about page. (Neither of their templates have been written
yet, though.)

core.views is now a directory rather than a single file, with
basic views (index and about) in one file and plain-text-generating
views in another.