Update navbar and index page for login
Navbar now includes username and logout link when a user is logged in.

Index page now has a Github login link if the user is logged out,
and only displays the links to visions if the user is logged in.

example.html has been deleted because it was just used as part of the
initial auth demo.