Add login_required decorator to plain text endpoints
In order to make this work we also need to add a custom
template for account/login.html.

This is a good temporary fix, I think, but it would make more sense
to default most of the website to requiring login and only make
a few pages available if you're logged out.